pyproject.toml
[project]
name = "bot2mvp"
version = "0.1.0"
description = "Telegram bot para alertas personalizadas de vacantes freelance"
requires-python = ">=3.10"

dependencies = [
    # Core Bot
    "python-telegram-bot>=20.0",
    # Data Validation & Serialization
    "pydantic>=2.0",
    # AI & LLM
    "langchain>=0.1",
    "langchain-google-genai>=0.0.15",
    "google-generativeai>=0.3",
    # API & Async
    "requests>=2.31",
    "aiohttp>=3.8",
    # Scheduling
    "apscheduler>=3.10",
    # Configuration
    "python-dotenv>=1.0",
    # Google Sheets (Database)
    "google-auth-oauthlib>=1.0",
    "google-auth-httplib2>=0.2",
    "google-api-python-client>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]
